+++ app/graphql/resolvers/projects_resolver.rb
@@ -12,6 +12,9 @@ module Resolvers
     option(:priorities, type: [String]) { |scope, value| scope.where(priority: value) }
     option(:construction_types, type: [String]) { |scope, value| scope.where(construction_type: value) }
     option(:internal_ids, type: [String]) { |scope, value| scope.where(internal_id: value) }
+    option(:customer_requests, type: [Boolean]) { |scope, value| scope.where(customer_request: value) }
+    option(:lot_numbers, type: [String]) { |scope, value| scope.where(lot_number: value) }
+    option(:kam_regions, type: [String]) { |scope, value| scope.where(kam_region: value) }
 
     option :buildings_count, type: [Int], with: :apply_buildings_filter, description: <<~DESC
       Send min and max values in the array. eg. [2, 10]. Only the first two values will be picked.