+++ app/models/concerns/hooks/project.rb
@@ -12,6 +12,11 @@ module Hooks
       after_save :update_projects_list, :update_users_list
       after_create :create_default_label_group, :update_status
       after_destroy :update_projects_list, :update_users_list
+
+      after_discard do
+        address_books.discard_all
+        buildings.discard_all
+      end
     end
 
     private