+++ db/migrate/20211020111623_add_sfn_to_competition.rb
@@ -4,7 +4,9 @@ class AddSfnToCompetition < ActiveRecord::Migration[6.1]
   disable_ddl_transaction!
 
   def change
-    add_column :admin_toolkit_competitions, :sfn, :boolean, default: false
-    add_index :admin_toolkit_competitions, :sfn, algorithm: :concurrently
+    safety_assured do
+      add_column :admin_toolkit_competitions, :sfn, :boolean, default: false
+      add_index :admin_toolkit_competitions, :sfn, algorithm: :concurrently
+    end
   end
 end