+++ db/migrate/20211125113806_add_discarded_at_to_buildings.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class AddDiscardedAtToBuildings < ActiveRecord::Migration[6.1]
+  disable_ddl_transaction!
+
+  def change
+    add_column :projects_buildings, :discarded_at, :datetime
+    add_index :projects_buildings, :discarded_at, algorithm: :concurrently
+  end
+end