+++ db/migrate/20211125113858_add_discarded_at_to_address_books.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class AddDiscardedAtToAddressBooks < ActiveRecord::Migration[6.1]
+  disable_ddl_transaction!
+
+  def change
+    add_column :projects_address_books, :discarded_at, :datetime
+    add_index :projects_address_books, :discarded_at, algorithm: :concurrently
+  end
+end