+++ db/migrate/20211125115233_update_projects_lists_to_version_3.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class UpdateProjectsListsToVersion3 < ActiveRecord::Migration[6.1]
+  def change
+    update_view :projects_lists, version: 3, revert_to_version: 2, materialized: true
+  end
+end